# ma-sharvari-ki-jai/server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ma Sharvari Ki Jai CRM API
  version: 0.2.0
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Customer:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        totalSpend: { type: number }
        visitCount: { type: integer }
        lastOrderDate: { type: string, format: date-time }
        tags: { type: array, items: { type: string } }
        attributes: { type: object }
        externalCustomerId: { type: string, description: 'Idempotency key from source system' }
        createdBy: { type: string, description: 'User owner (server-managed)' }
...
paths:
...
  /api/customers:
    post:
      summary: Ingest customers
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Customer'
                - type: array
                  items: { $ref: '#/components/schemas/Customer' }
      description: |
        Idempotent ingest by `externalCustomerId` per user. Re-sending the same payload updates the record.
      responses:
        '201': { description: Created }
    get:
      summary: List customers
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Search in name/email
        - in: query
          name: email
          schema: { type: string }
        - in: query
          name: tags
          schema: { type: string }
          description: Comma-separated list; matches customers containing all tags
        - in: query
          name: minSpend
          schema: { type: number }
        - in: query
          name: maxSpend
          schema: { type: number }
        - in: query
          name: dateFrom
          schema: { type: string, format: date-time }
        - in: query
          name: dateTo
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: OK }
...